requirements.txt
# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
python-multipart>=0.0.6

# ML/AI
torch>=2.0.0
transformers>=4.35.0
huggingface-hub>=0.21.0
accelerate>=0.25.0

# Visualization
plotly>=5.18.0

# Data handling
pyyaml>=6.0
aiofiles>=23.2.0

# Note: HatCat is installed separately from the sibling directory
